test(config): reuse shared setup
